requirements.txt
Flask==2.3.3
google-cloud-translate==3.11.4
requests==2.31.0
python-dotenv==1.0.0
azure-cognitiveservices-speech==1.32.1
azure-identity==1.15.0
Werkzeug==2.3.7
scikit-learn==1.3.0
numpy==1.24.3
kaggle==1.6.14